Check 2
Created basic inventory mechanics. Can look at a GameObject, press E and put it in inventory. Needs to be debugged, as well as to create a UI.
Created an editor tool to quickly make icons for Prefabs (they will be used for inventory)
Created a world to be used in the resulting game.